refactor: move arquivo de montagem do socketIo para pasta de lib
